readCallback may be called on EOF
Due to a limitation in PHP, the code in readCallback may be called when
dealing with network sockets and EOF is received (ie. the peer disconnected).

To work around this, when fread() is called in readCallback and the stream
is actually unusable due to EOF/disconnection, a special exception is thrown.
This exception is caught by the manager and closeCallback will be called.